// JwtWork/ClientApp/src/fragments/rndclktypes.d.ts
export interface IKnotCore {
  // The radius of the pointer in SVG units.
  // Default value = 10.
  radius?: number
  // The width of the border around the pointer in SVG units.
  // Default value = 0.
  border?: number
}

export interface IKnotProps extends IKnotCore {
  // The value associated with the pointer.
  // This value determines the position of the pointer on the slider.
  // Default value = 0.
  value?: string | number

  // The background color of the pointer.
  // Default value = #163a86.
  bgColor?: string

  // The background color of the pointer when it is selected or active.
  // Default value = #000.
  bgColorSelected?: string

  // The background color of the pointer when it is disabled.
  // Default value = #a8a8a8.
  bgColorDisabled?: string

  // The background color of the pointer when it is hovered.
  // Default value is the same as bgColor.
  bgColorHover?: string

  // The color of the border around the pointer.
  // Default value = #000.
  borderColor?: string

  // Specifies whether the pointer is disabled or not.
  // If set to true, the pointer will be inactive and non-interactive.
  // Default value = false.
  disabled?: boolean

  // A string that describes the purpose or function of the pointer for accessibility purposes.
  // This label will be used as the aria-label attribute for the pointer element.
  // Default value = undefined.
  ariaLabel?: string
}
interface IAngleInfo {
  angleDeg: number
  index: number
}
interface KnotState {
  knots: IAngleInfo[]
  count: number
  angleShift: number
  
  push: (newknot: IAngleInfo, setCycle: React.Dispatch<React.SetStateAction<number>>) => void
  peek: (index: number) => IAngleInfo|null
  isInbound: (newknot: IAngleInfo) => boolean
  noMove: (newknot: IAngleInfo) => boolean
  getNewIndex: () => number
  setShiftOnce: (shift:number) => void  

}

export interface IKnotInstance extends Omit<Required<IKnotProps>, 'value|ariaLabel'>,IAngleInfo {
  id: string
  prevAngleDeg: number
  value?: string | number
  ariaLabel?: string
}
export interface IKnotBagInstance {
  knots: IKnotInstance[]
  maxRadius: number
}
export interface IStrokeProps {
  strokeDasharray: string
  strokeOffset: number
}
// ...
export interface IClockCore {
  cx: number
  cy: number
  radius: number
  startAngleDeg: number
  endAngleDeg: number
  clockAngleShift: number
}